add a startup probe to the hogarama container
the liveness probe had an initial delay which is not ideal since the startup time might not always be the same.
The startup probe probes for liveness at the start of the container. It does so a maximum of 20 times every 10 seconds before failing.
After the startup probe gets a positive return (200) the liveness probe takes over.